requirements.txt
Flask~=2.3.2
Flask-PyMongo~=2.3.0
Flask-JWT-Extended~=4.5.0
python-dotenv~=1.0.0
from config import Config